Read streaming fees from on-chain via set.js (#380)
* WIP: initial set.js impl

`yarn start` works now. Had to build set.js locally with
SetProtocol PR 76. So this PR will be blocked on upgrading to a release
of set.js with that change.

* WIP: fetch quantity of DPI component from on-chain

- Add ethers as a dep
- Need `resolutions` block to resolve error

* Remove unnecessary provider in favor of existing provider

* Fix indentaiton

Prettier should really handle this for us

* Upgrade to set.js 0.3.0

* Pass moduleAddresses

* Revert DpiIndexComponentsProvider

* WIP: fetch streaming fees from set.js

* Fetch streaming fees on-chain

I had to update the contract addresses in .env and .env.prod to get this
to work

* Fetch all product streaming fees from on-chain

* Rename context

* Address @0xModene comments

* Prefer absolute imports