Read set component quantities from on-chain
This is a V0 of reading DPI index components from on-chain using set.js
There is room for a refactor here. Instead of plumbing two types of
components (`IndexComponents` and `SetComponents`) into child React
components, we can collapse these two objects at the provider / data
fetching layer. This will enable us to unit test the data merging logic
which is the sketchiest part about this PR.

We need to fallback to the existing `IndexComponents` fetched via Set
API because it contains other component metadata (e.g. image, symbol,
dailyPercentChange, etc.) which is used in the table.

Additionally we want to fallback to the existing `IndexComponents`
because the set.js request will only work after a user has connected
their wallet. We want the DPI page to still contain index component
information if a user hasn't connected their wallet yet.